Finished Low Bar Skeleton
I finished the skeleton of the low bar code by adding some code that
will have the robot launch the boulder and cross back under the low bar
into the neutral zone. The robot will also send a message to the driver
station whenever it shifts gears and completes parts of the autonomous.